separated code into driver and model scripts
